+++ app/components/graph/graph-directives.js
@@ -428,3 +428,34 @@ angular.module('lizard-nxt')
   };
 
 }]);
+
+
+/**
+ * Creates a specific line and point graph for levee crosssections.
+ *
+ * Content should contain property line for elevation data and the property
+ * points for timeseries values of wells.
+ */
+angular.module('lizard-nxt')
+.directive('crossSection', [function () {
+  var link = function (scope, element, attrs, graphCtrl) {
+
+    var content = graphCtrl.content,
+        graph = graphCtrl.graph;
+
+    graph.drawCrosssection(content);
+
+    // Function to call when data changes
+    graphCtrl.updateData = graph.drawCrosssection;
+
+    };
+
+    return {
+      require: 'graph',
+      link: link,
+      restrict: 'A'
+    };
+
+  }
+]);
+