+++ CHANGES.rst
@@ -3,7 +3,7 @@ Changelog of lizard-nxt client
 
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
--
+- Add share favourites.
 
 
 Release 2.9.4 (2016-3-1)